Extract the null handling logic to a superclass AbstractBinaryExpression
